fix wrangler pipeline test
The wrangler pipeline test does not require the wrangler
service to be running, so removed the part that deploys and starts
the service. This was failing due to the fact that wrangler must
now be deployed to the system namespace.

Also fixed a compilation issue due to a class being moved.