CDAP-16429 log a warning if pipeline cleanup fails
The pubsub test was flaky because it would occasionally get a 400
error that the pipeline is not running when the stop call is made.
Changed the test to just log a warning in this scenario since the
test was actually successful. CDAP may have some underlying bugs
causing this behavior though.